+++ package.json
@@ -7,6 +7,7 @@
   "private": false,
   "scripts": {
     "build": "tsc --project ./",
+    "setup": "tsnd setup.js && install",
     "prestart": "tsnd prestart.ts",
     "start": "tsnd src/server.ts"
   },
